+++ src/deprecated/pipe/__test__/SpreadPipe-test.js
@@ -1,7 +1,7 @@
 import test from 'ava';
 import {shallow} from 'enzyme';
 import React from 'react';
-import SpreadPipe from './SpreadPipe';
+import SpreadPipe from '../SpreadPipe';
 import {Map} from 'immutable';
 import sinon from 'sinon';
