+++ src/util/__test__/SpruceComponent-test.js
@@ -1,7 +1,7 @@
 import test from 'ava';
 import React from 'react';
 import {shallow} from 'enzyme';
-import SpruceComponent from './SpruceComponent';
+import SpruceComponent from '../SpruceComponent';
